third_party/WebKit/LayoutTests/fast/dom/HTMLTemplateElement/custom-element-wrapper-gc-expected.txt

Custom element wrappers in templates should be retained

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing major GC...
PASS customElement.expando is "present"
Testing minor GC...
PASS customElement.expando is "present"
PASS successfullyParsed is true

TEST COMPLETE
